Backend/flaskapp/requirements.txt
Flask>=2.0.2
Flask-RESTful>=0.3.9
flask-mongoengine>=1.0.0
requests>=2.26.0
fastapi>=0.70.0
argon2-cffi>=20.1.0
Flask-JWT-Extended>=4.3.1